+++ bin/read.js
@@ -15,8 +15,8 @@ var axios_1 = __importDefault(require("axios"));
 var js_yaml_1 = __importDefault(require("js-yaml"));
 var fs_1 = __importDefault(require("fs"));
 var path_1 = __importDefault(require("path"));
-var T = __importStar(require("fp-ts/lib/Task"));
-var E = __importStar(require("fp-ts/lib/Either"));
+var IE = __importStar(require("fp-ts/lib/IOEither"));
+var TE = __importStar(require("fp-ts/lib/TaskEither"));
 var pipeable_1 = require("fp-ts/lib/pipeable");
 var utils_1 = require("./utils");
 var function_1 = require("fp-ts/lib/function");
@@ -24,48 +24,34 @@ function isUrl(pathOrUrl) {
     return pathOrUrl.indexOf("https://") !== -1;
 }
 function getContentFromURL(url) {
-    return function () {
-        return axios_1.default({
-            method: "get",
-            url: url
-        })
-            .then(function (_a) {
+    return TE.tryCatch(function () {
+        return axios_1.default({ method: "get", url: url }).then(function (_a) {
             var data = _a.data;
-            return E.right(typeof data === "object" ? data : js_yaml_1.default.safeLoad(data));
-        })
-            .catch(function (error) { return E.left(error); });
-    };
+            return typeof data === "object" ? data : js_yaml_1.default.safeLoad(data);
+        });
+    }, function (error) { return error; });
 }
 function getContentFromPath(file) {
-    return function () {
-        try {
-            var ext = path_1.default.extname(file);
-            var content = fs_1.default.readFileSync(file, "utf8");
-            var swagger = ext === ".yaml" || ext === ".yml"
-                ? js_yaml_1.default.safeLoad(content)
-                : JSON.parse(content);
-            return E.right(swagger);
-        }
-        catch (error) {
-            return E.left(error);
-        }
-    };
+    return IE.tryCatch(function () {
+        var ext = path_1.default.extname(file);
+        var content = fs_1.default.readFileSync(file, "utf8");
+        var swagger = ext === ".yaml" || ext === ".yml"
+            ? js_yaml_1.default.safeLoad(content)
+            : JSON.parse(content);
+        return swagger;
+    }, function (error) { return error; });
 }
 function getContent(source) {
-    return pipeable_1.pipe(source, utils_1.doIf(isUrl, function (source) { return getContentFromURL(source); }, function (source) { return T.fromIO(getContentFromPath(source)); }));
+    return pipeable_1.pipe(source, utils_1.doIfElse(isUrl, function (source) { return getContentFromURL(source); }, function (source) { return TE.fromIOEither(getContentFromPath(source)); }));
 }
 function patchSwagger(patchSource) {
     return function (swagger) {
-        return pipeable_1.pipe(patchSource, function (source) { return getContent(source); }, T.map(E.map(utils_1.patch(swagger))));
+        return pipeable_1.pipe(patchSource, function (source) { return getContent(source); }, TE.map(utils_1.patch(swagger)));
     };
 }
 function getSwagger(patchSource) {
     return function (source) {
-        return pipeable_1.pipe(source, function (source) { return getContent(source); }, utils_1.doIf(function () { return patchSource != null; }, function (task) {
-            return pipeable_1.pipe(task, T.chain(function (eitherSwagger) {
-                return E.either.traverse(T.task)(eitherSwagger, patchSwagger(patchSource));
-            }), T.map(E.flatten));
-        }, function_1.identity));
+        return pipeable_1.pipe(source, function (source) { return getContent(source); }, utils_1.doIf(function_1.constant(patchSource != null), TE.chain(patchSwagger(patchSource))));
     };
 }
 exports.getSwagger = getSwagger;